refactored search related code into its own local component
